ai_hint_project/src/ai_hint_project/requirements.txt
streamlit
crewai
crewai-tools
pyyaml
faiss-cpu
sentence-transformers
langchain
langchain-community
numpy
scikit-learn
groq
langchain-groq
